+++ CHANGES.rst
@@ -2,7 +2,7 @@
 Cython Changelog
 ================
 
-3.3.0b1 (2026-07-2?)
+3.3.0b1 (2026-08-??)
 ====================
 
 Features added
@@ -32,6 +32,9 @@ Features added
   use direct C access in CPython.
   (Github issue :issue:`7783`)
 
+* Annotations on global variables are now used by type inference.
+  (Github issue :issue:`7877`)
+
 * C property setters can now explicitly propagate exceptions, instead of always triggering
   a call to ``PyErr_Occurred()`` by returning ``void``.
   (Github issue :issue:`7791`)
@@ -46,6 +49,9 @@ Features added
 * The typed tuple syntax ``tuple[atype, ...]`` for homogeneous tuples is supported.
   (Github issue :issue:`7798`)
 
+* The return type inference for calls to builtin methods was improved, including dict views.
+  (Github issues :issue:`7887`, :issue:`7888`)
+
 * ``cython.py_int`` (and the same for ``py_float``, ``py_complex`` and ``py_bool``)
   can be used to refer to Python's builtin types in a C type context, e.g. after ``cdef``,
   where they are normally shadowed by the C types of the same name.
@@ -55,8 +61,14 @@ Features added
   (similar to constant tuple and slice objects).
   Original Patch by Zhenbo Li. (Github issue :issue:`2741`)
 
+* Formatting C floating point values in f-strings is faster.
+  Patch by Vladimir Saraikin.  (Github issue :issue:`7797`)
+
 * ``bytearray.extend(bytes)`` is faster.
   (Github issue :issue:`7797`)
+
+* Single character `in`-tests on ``str``, ``bytes`` and ``bytearray`` are optimised.
+  (Github issue :issue:`3888`)
 
 * ``assert`` conditions are constant-folded.
   (Github issue :issue:`7797`)
@@ -102,7 +114,7 @@ Bugs fixed
 
 * Casting to an unresolved ``typeof()`` type (e.g. ``cython.cast(cython.typeof(x), ...)``
   where the type could not be inferred) crashed the compiler instead of reporting an error.
-  Patch by vsaraikin.  (Github issue :issue:`7683`)
+  Patch by Vladimir Saraikin.  (Github issue :issue:`7683`)
 
 * ``cpdef fused`` functions generated redundant code.
   (Github issue :issue:`7778`)
@@ -122,7 +134,7 @@ Bugs fixed
 
 * Iterating over a container with an incorrectly declared item type could generate
   incorrect C code in 3.3.0a1.
-  (Github issue :issue:`7775`)
+  (Github issues :issue:`7775`, :issue:`7889`)
 
 * ``NULL`` pointer comparisons could fail to compile with C++ in 3.3.0a1.
   Patch by Vyas Ramasubramani.  (Github issue :issue:`7766`)
@@ -151,6 +163,9 @@ Other changes
   due to lack of support in the runtimes, but is intended for initial testing
   and future improvements.
   (Github issue :issue:`7831`)
+
+* Cython wheels now use a shared utility code module to reduce their size.
+  (Github issue :issue:`7865`)
 
 
 3.3.0a1 (2026-06-24)